Refactoring: about file navigator
