+++ index.html
@@ -17,9 +17,7 @@
               <span class="icon icon-reply" id="move-parent-directory"></span>
             </span>
             <h5 class="nav-group-title">Directory</h5>
-            <span class="nav-group-item" id="path-directory">
-              <span class="icon icon-folder"></span>
-              <span id="path-directory-inner"></span>
+            <span class="nav-group-item nav-group-item-disactive" id="path-directory" >
             </span>
             <h5 class="nav-group-title">Files</h5>
             <nav id="directory-tree"></nav>
