+++ lib/define.js
@@ -1,0 +1,1 @@
+module.exports.rootPath = process.env[process.platform == "win32" ? "USERPROFILE" : "HOME"];
