+++ lib/ui.js
@@ -5,8 +5,9 @@
 const directoryTree = document.getElementById('directory-tree')
 module.exports.directoryTree = directoryTree
 module.exports.sideBar = document.getElementById('sidebar')
-module.exports.directoryContent = document.getElementById('path-directory')
-module.exports.directoryContentInner = document.getElementById('path-directory-inner')
+
+const directoryPath = document.getElementById('path-directory')
+module.exports.directoryPath = directoryPath
 
 // get element
